refactor(web): extract guild faction helpers and rename resolveSourceLabel variants
- Create apps/web/src/lib/guild-faction.ts with ReportGuildFaction type,
  normalizeGuildFaction, and resolveTitleRowClass — previously duplicated
  verbatim across 4 sibling report-list components
- Rename ambiguous resolveSourceLabel in each file to a descriptive name:
  resolveHostPrefixLabel (host-string based) and resolveAccountSourceLabel
  (guild/personal source type based)
