+++ apps/web/src/components/recent-reports-list.tsx
@@ -6,6 +6,11 @@ import type { FC } from 'react'
 import { Link } from 'react-router-dom'
 
 import { formatReportHeaderDate } from '../lib/format'
+import {
+  type ReportGuildFaction,
+  normalizeGuildFaction,
+  resolveTitleRowClass,
+} from '../lib/guild-faction'
 import { cn } from '../lib/utils'
 import type {
   ExampleReportLink,
@@ -27,38 +32,7 @@ export type RecentReportsListProps = {
   ) => void
 }
 
-type ReportGuildFaction = 'alliance' | 'horde' | null
-
-function normalizeGuildFaction(
-  value: string | null | undefined,
-): ReportGuildFaction {
-  if (!value) {
-    return null
-  }
-
-  const normalized = value.trim().toLowerCase()
-  if (normalized === 'alliance') {
-    return 'alliance'
-  }
-  if (normalized === 'horde') {
-    return 'horde'
-  }
-
-  return null
-}
-
-function resolveTitleRowClass(faction: ReportGuildFaction): string {
-  if (faction === 'alliance') {
-    return 'text-sky-600 dark:text-sky-400'
-  }
-  if (faction === 'horde') {
-    return 'text-red-600 dark:text-red-400'
-  }
-
-  return ''
-}
-
-function resolveSourceLabel(report: RecentReportEntry): string {
+function resolveHostPrefixLabel(report: RecentReportEntry): string {
   const hostPrefix = report.sourceHost.split('.')[0]?.toLowerCase()
 
   return hostPrefix ?? 'unknown'
@@ -97,7 +71,7 @@ export const RecentReportsList: FC<RecentReportsListProps> = ({
         const isInaccessible = report.isAccessible === false
         const isDisabled = isArchived || isInaccessible
         const guildName = report.guildName ? `<${report.guildName}>` : null
-        const sourceLabel = resolveSourceLabel(report)
+        const sourceLabel = resolveHostPrefixLabel(report)
         const titleParts = [report.title || report.reportId, guildName]
           .filter((value): value is string => Boolean(value))
           .join(' ')
